+++ weapon_tweaker/scripts/mods/weapon_tweaker/weapon_tweaker_localization.lua
@@ -1339,6 +1339,9 @@ local _cwv_variant_labels = {
     cwv_wh_dual_axes = "Saltzpyre: Dual Axes",
     cwv_es_dual_maces = "Kruber: Dual Maces",
     cwv_wh_dual_maces = "Saltzpyre: Dual Maces",
+    cwv_dr_dawi_mace = "Bardin: Dawi Mace",
+    cwv_dr_dawi_mace_shield = "Bardin: Dawi Mace and Shield",
+    cwv_dr_dawi_dual_maces = "Bardin: Dawi Dual Maces",
     cwv_es_dual_warpriest_hammers = "Kruber: Dual Warrior-Priest Hammers",
     cwv_es_warpriest_hammer_shield = "Kruber: Warrior-Priest Hammer and Shield",
 }